chore (ai): extract ResponseMessage to separate file
