+++ packages/ai/core/generate-text/generate-text.ts
@@ -34,7 +34,8 @@ import { GenerateTextResult } from './generate-text-result';
 import { Output } from './output';
 import { parseToolCall } from './parse-tool-call';
 import { asReasoningText } from './reasoning';
-import { ResponseMessage, StepResult } from './step-result';
+import { ResponseMessage } from './response-message';
+import { StepResult } from './step-result';
 import { toResponseMessages } from './to-response-messages';
 import { ToolCallArray } from './tool-call';
 import { ToolCallRepairFunction } from './tool-call-repair';
