+++ packages/ai/core/generate-text/response-message.ts
@@ -1,0 +1,12 @@
+import { AssistantModelMessage, ToolModelMessage } from '../prompt/message';
+
+/**
+A message that was generated during the generation process.
+It can be either an assistant message or a tool message.
+ */
+export type ResponseMessage = (AssistantModelMessage | ToolModelMessage) & {
+  /**
+  Message ID generated by the AI SDK.
+     */
+  id: string;
+};
