+++ packages/ai/core/generate-text/step-result.ts
@@ -1,5 +1,4 @@
 import { ReasoningPart } from '../prompt/content-part';
-import { AssistantModelMessage, ToolModelMessage } from '../prompt/message';
 import {
   CallWarning,
   FinishReason,
@@ -11,20 +10,10 @@ import { Source } from '../types/language-model';
 import { LanguageModelUsage } from '../types/usage';
 import { ContentPart } from './content-part';
 import { GeneratedFile } from './generated-file';
+import { ResponseMessage } from './response-message';
 import { ToolCallArray } from './tool-call';
 import { ToolResultArray } from './tool-result';
 import { ToolSet } from './tool-set';
-
-/**
-A message that was generated during the generation process.
-It can be either an assistant message or a tool message.
- */
-export type ResponseMessage = (AssistantModelMessage | ToolModelMessage) & {
-  /**
-Message ID generated by the AI SDK.
-   */
-  id: string;
-};
 
 /**
  * The result of a single step in the generation process.
