+++ packages/ai/core/generate-text/stream-text-result.ts
@@ -13,10 +13,11 @@ import { LanguageModelResponseMetadata } from '../types/language-model-response-
 import { LanguageModelUsage } from '../types/usage';
 import { ContentPart } from './content-part';
 import { GeneratedFile } from './generated-file';
-import { ResponseMessage, StepResult } from './step-result';
+import { StepResult } from './step-result';
 import { ToolCallUnion } from './tool-call';
 import { ToolResultUnion } from './tool-result';
 import { ToolSet } from './tool-set';
+import { ResponseMessage } from './response-message';
 
 export type DataStreamOptions = {
   /**
